Split light fixture from implementation to only import when fixture is actually used
