+++ tests/components/conftest.py
@@ -1,10 +1,13 @@
 """Fixtures for component testing."""
 
 from collections.abc import Generator
-from typing import Any
+from typing import TYPE_CHECKING, Any
 from unittest.mock import MagicMock, patch
 
 import pytest
+
+if TYPE_CHECKING:
+    from tests.components.light.common import MockLight, SetupLightPlatformCallable
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -101,3 +104,25 @@ def prevent_ffmpeg_subprocess() -> Generator[None, None, None]:
         "homeassistant.components.ffmpeg.FFVersion.get_version", return_value="6.0"
     ):
         yield
+
+
+@pytest.fixture
+def mock_light_entities() -> list["MockLight"]:
+    """Return mocked light entities."""
+    from homeassistant.const import STATE_OFF, STATE_ON
+
+    from tests.components.light.common import MockLight
+
+    return [
+        MockLight("Ceiling", STATE_ON),
+        MockLight("Ceiling", STATE_OFF),
+        MockLight(None, STATE_OFF),
+    ]
+
+
+@pytest.fixture
+def setup_light_platform() -> "SetupLightPlatformCallable":
+    """Return a callable to set up the mock light entity component."""
+    from tests.components.light.common import setup_light_platform
+
+    return setup_light_platform
